+++ ipspot/params.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 """ipspot params."""
 from enum import Enum
+import re
 
 IPSPOT_VERSION = "0.2"
 
@@ -16,6 +17,8 @@ REQUEST_HEADERS = {
     'User-Agent': 'IPSpot/{version} ({repo})'.format(version=IPSPOT_VERSION, repo=IPSPOT_REPO),
     'Accept': 'application/json'
 }
+
+IPV4_REGEX = re.compile(r'^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$')
 
 
 class IPv4API(Enum):